Extract Configuration label text to a resource. Add missing file headers.
