+++ GoogleCloudExtension/GoogleCloudExtension/PublishDialog/Steps/Gke/GkeStepContent.xaml
@@ -4,7 +4,6 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:local="clr-namespace:GoogleCloudExtension.PublishDialog.Steps.Gke"
-             xmlns:ext="clr-namespace:GoogleCloudExtension"
              xmlns:utils="clr-namespace:GoogleCloudExtension.Utils;assembly=GoogleCloudExtension.Utils"
              xmlns:steps="clr-namespace:GoogleCloudExtension.PublishDialog.Steps"
              mc:Ignorable="d"
@@ -15,7 +14,7 @@
     </UserControl.Resources>
 
     <StackPanel>
-        <steps:PublishDialogStepHeaderContent/>
+        <steps:PublishDialogStepHeaderControl/>
 
         <!-- The message that the project is being validated. -->
         <steps:LoadingProjectDataControl
