--- GoogleCloudExtension/GoogleCloudExtension/PublishDialog/Steps/PublishDialogStepHeaderContent.xaml.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace GoogleCloudExtension.PublishDialog.Steps
-{
-    /// <summary>
-    /// Interaction logic for PublishDialogStepHeaderContent.xaml
-    /// </summary>
-    public partial class PublishDialogStepHeaderContent : UserControl
-    {
-        public PublishDialogStepHeaderContent()
-        {
-            InitializeComponent();
-        }
-    }
-}
